Move to queries folder
