+++ app/auth/queries/getSignature.ts
